# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Description of my API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/create_promo_code": {
            "post": {
                "tags": [
                    "Admin Promo Codes"
                ],
                "summary": "Create a new promo code",
                "description": "Creates a new promo code with the specified details.",
                "operationId": "6d005de21dd00cde42a2776937bdede4",
                "requestBody": {
                    "description": "Promo code details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "agent_id": {
                                        "description": "ID of the agent associated with the promo code",
                                        "type": "integer"
                                    },
                                    "code": {
                                        "description": "Promo code (min: 5 characters)",
                                        "type": "string"
                                    },
                                    "end_date": {
                                        "description": "End date of the promo code (format: YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "job_count": {
                                        "description": "Number of jobs for the promo code to be valid",
                                        "type": "integer"
                                    },
                                    "discount": {
                                        "description": "Discount amount for the promo code",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Promo code created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Promo code created successfully",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/single_page_promo_code": {
            "get": {
                "tags": [
                    "Admin Promo Codes"
                ],
                "summary": "Get information of a single promo code",
                "description": "Retrieves information of a single promo code based on the provided code_id.",
                "operationId": "8dd11548b350cb1632e5b80512c7378b",
                "parameters": [
                    {
                        "name": "code_id",
                        "in": "query",
                        "description": "ID of the promo code to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promo code information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Promo code not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Code Not found",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/update_promo_code": {
            "post": {
                "tags": [
                    "Admin Promo Codes"
                ],
                "summary": "Update promo code information",
                "description": "Updates the information of an existing promo code based on the provided code_id.",
                "operationId": "5205f110999e76c734228b1c8f0ac556",
                "requestBody": {
                    "description": "Promo code details to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code_id": {
                                        "description": "ID of the promo code to update",
                                        "type": "integer"
                                    },
                                    "agent_id": {
                                        "description": "ID of the agent associated with the promo code",
                                        "type": "integer"
                                    },
                                    "code": {
                                        "description": "Promo code (min: 5 characters)",
                                        "type": "string"
                                    },
                                    "end_date": {
                                        "description": "End date of the promo code (format: YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "job_count": {
                                        "description": "Number of jobs for the promo code to be valid",
                                        "type": "integer"
                                    },
                                    "discount": {
                                        "description": "Discount amount for the promo code",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Promo code information updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Promo code updated successfully",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/get_all_promo_codes": {
            "get": {
                "tags": [
                    "Admin Promo Codes"
                ],
                "summary": "Get all promo codes with optional search",
                "description": "Retrieves a paginated list of all promo codes with optional search functionality.",
                "operationId": "cc082607e0f39e02c3632ef5db7d607c",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search keyword to filter promo codes by code",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promo codes retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    }
}